requirements.txt
# Core Django dependencies
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
django-environ>=0.10.0

# Database
psycopg2-binary>=2.9.0

# Instagram API
instagrapi>=1.19.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0

# Logging
logging>=0.4.9.6

# LLM
llama_index==0.12.40
Flask==3.1.1
hmmlearn==0.3.3

# Mongodb
pymongo==4.13.0

# Env loader
python-dotenv==1.1.0
environ==1.0

# Numpy 
numpy==1.23.5

# JWT
PyJWT==2.10.1
rest_framework_simplejwt==0.0.2

# Requests
Requests==2.32.3
